Move homepage to App router and tidy up shared layouts
