+++ apps/web/src/components/JoinTheCommunity/JoinTheCommunity.tsx
@@ -26,7 +26,7 @@ const partners = [
   'heroes-of-mavia',
 ] as const;
 
-export function JoinTheCommunity() {
+export async function JoinTheCommunity() {
   return (
     <div className="flex w-full max-w-[780px] flex-col justify-center gap-8 bg-black px-6 py-12 md:gap-6 lg:justify-start lg:px-7 xl:max-w-[1440px]">
       <h2 className="font-display text-3xl text-white md:text-6xl xl:max-w-[50%]">
