+++ apps/web/src/components/Layout/Footer/Footer.tsx
@@ -1,3 +1,4 @@
+'use client';
 import { useState, useCallback } from 'react';
 import Link from 'next/link';
 import { CookiePreferencesModal } from '@coinbase/cookie-banner';
