+++ apps/web/src/components/Layout/Nav/Nav.tsx
@@ -1,18 +1,24 @@
+'use client';
 import Link from 'next/link';
-import { useRouter } from 'next/router';
 
 import { Logo } from '../../Logo/Logo';
 import DesktopNav from './DesktopNav';
 import MobileMenu from './MobileMenu';
 import Banner from 'base-ui/components/Layout/Nav/Banner';
+import { usePathname } from 'next/dist/client/components/navigation';
 
-type NavProps = {
-  color: 'white' | 'black';
-};
+const BLACK_NAV_PATHS = [
+  '/',
+  '/jobs/apply',
+  '/cookie-policy',
+  '/third-party-cookies',
+  '/onchainsummer',
+  '/name',
+];
 
-export function Nav({ color }: NavProps) {
-  const { pathname } = useRouter();
-
+export function Nav() {
+  const pathname = usePathname();
+  const color = pathname && BLACK_NAV_PATHS.includes(pathname) ? 'black' : 'white';
   return (
     <>
       <Banner
@@ -20,9 +26,9 @@ export function Nav({ color }: NavProps) {
         href="https://onchainkit.xyz/?utm_source=basedotorg&utm_medium=banner"
         text="Build on Base in minutes with OnchainKit!"
       />
-      <nav className="bg-transparent z-10 flex h-24 w-full max-w-[1440px] flex-row items-center justify-between gap-16 self-center p-8">
+      <nav className="z-10 flex h-24 w-full max-w-[1440px] flex-row items-center justify-between gap-16 self-center bg-transparent p-8">
         <Link href="/" aria-label="Base Homepage">
-          <Logo color={color} path={pathname} width="106px" />
+          <Logo color={color} width="106px" />
         </Link>
         <DesktopNav color={color} />
         <MobileMenu color={color} />
