+++ apps/web/src/components/Partnerships/Partnerships.tsx
@@ -20,7 +20,7 @@ function LearnMoreButton({ ariaLabel }: LearnMoreButtonProps) {
   );
 }
 
-export function Partnerships() {
+export async function Partnerships() {
   return (
     <div className="flex w-full max-w-[1440px] flex-row flex-wrap items-start justify-between bg-black px-8 py-12">
       <div className="flex w-full flex-col space-y-8 pb-16 lg:basis-1/2 lg:space-y-12 lg:pt-1">
